split tests to ci/!ci
